+++ src/libraries/System.Security.AccessControl/src/System/Security/Policy/Evidence.cs
@@ -1,6 +1,8 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+#nullable enable
+
 using System.Collections;
 
 namespace System.Security.Policy
