+++ src/libraries/System.Threading.AccessControl/ref/System.Threading.AccessControl.csproj
@@ -8,8 +8,8 @@
     <Compile Include="System.Threading.AccessControl.cs" Condition="'$(TargetFramework)' != 'net461'" />
     <Compile Include="System.Threading.AccessControl.net461.cs" Condition="'$(TargetFramework)' == 'net461'" />
   </ItemGroup>
-  <ItemGroup Condition="'$(TargetFramework)' != 'net461'">
-    <ProjectReference Include="..\..\System.Security.AccessControl\ref\System.Security.AccessControl.csproj" />
-    <ProjectReference Include="..\..\System.Security.Principal.Windows\ref\System.Security.Principal.Windows.csproj" />
+  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
+    <PackageReference Include="System.Security.AccessControl" Version="$(SystemSecurityAccessControlVersion)" />
+    <PackageReference Include="System.Security.Principal.Windows" Version="$(SystemSecurityPrincipalWindowsVersion)" />
   </ItemGroup>
 </Project>
